Time: 29 ms (84.33%) | Memory: 101.6 MB (30.81%) - LeetSync

+++ leetcode/155-min-stack/min-stack.java
@@ -12,7 +12,7 @@ class MinStack {
        if(stack.isEmpty()){
         min=value;
        }
-       long diff=(long)value-min;
+       long diff=value-min;
        stack.push(diff);
        if(diff<0){
         min=value;
